refactor(gke): split version tests into version_test.go
Mirrors the gke.go → version.go production split. Pure logic tests
(ResolveVersionForChannel) are now separate from the HTTP/cache tests.
